update to 3.1.4
