+++ MercurySDK/MercurySDK.framework/Versions/3.1.4/Headers/MercuryAdHeader.h
@@ -1,0 +1,35 @@
+//
+//  BYMercuryAdHeader.h
+//  Example
+//
+//  Created by CherryKing on 2019/11/14.
+//  Copyright © 2019 CherryKing. All rights reserved.
+//
+
+#ifndef BYMercuryAdHeader_h
+#define BYMercuryAdHeader_h
+
+// MARK: ======================= MercuryBannerAdView =======================
+#import "MercuryBannerAdView.h"
+
+// MARK: ======================= MercurySplashAd =======================
+#import "MercurySplashAd.h"
+
+// MARK: ======================= MercuryInterstitialAd =======================
+#import "MercuryInterstitialAd.h"
+
+// MARK: ======================= MercuryRewardVideoAd =======================
+#import "MercuryRewardVideoAd.h"
+
+// MARK: ======================= MercuryNativeExpressAd =======================
+#import "MercuryNativeExpressAd.h"
+
+// MARK: ======================= MercuryPrerollAdView =======================
+#import "MercuryPrerollAdView.h"
+
+// MARK: ======================= MercuryNativeAd =======================
+#import "MercuryNativeAd.h"
+#import "MercuryNativeAdView.h"
+#import "MercuryNativeAdDataModel.h"
+
+#endif
