+++ MercurySDK/MercurySDK.framework/Versions/3.1.4/Headers/MercurySDK.h
@@ -1,0 +1,25 @@
+//
+//  MercurySDK.h
+//  MercurySDK
+//
+//  Created by 程立卿 on 2020/1/19.
+//  Copyright © 2020 Mercury. All rights reserved.
+//
+
+#import <Foundation/Foundation.h>
+
+#if __has_include(<MercurySDK/MercuryAdHeader.h>)
+#import <MercurySDK/MercuryAdHeader.h>
+#else
+#import "MercuryAdHeader.h"
+#endif
+#if __has_include(<MercurySDK/MercuryConfigManager.h>)
+#import <MercurySDK/MercuryConfigManager.h>
+#else
+#import "MercuryConfigManager.h"
+#endif
+#if __has_include(<MercurySDK/MercuryPubEnumHeader.h>)
+#import <MercurySDK/MercuryPubEnumHeader.h>
+#else
+#import "MercuryPubEnumHeader.h"
+#endif
