+++ MercurySDK/MercurySDK.framework/Versions/Current
@@ -1,1 +1,1 @@
-3.1.3.9
+3.1.4
